requirements.txt
alembic==1.17.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
facenet-pytorch==2.5.3
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
humanfriendly==10.0
idna==3.11
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
mako==1.3.10
markupsafe==3.0.3
matplotlib==3.10.7
ml-dtypes==0.5.4
mpmath==1.3.0
networkx==3.5
numpy==2.3.5
onnx==1.19.1
onnx-ir==0.1.12
onnxruntime==1.23.2
onnxscript==0.5.6
opencv-python==4.12.0.88
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sqlalchemy==2.0.44
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
